feat: LevelCli 关卡协作工具 + 分析增强 + 语义验证
新增 Match3.LevelCli 控制台项目，提供关卡设计的 validate/analyze/create 命令闭环：
- validate: 结构验证 + ObjectiveConsistencyChecker 语义检测（目标层不存在、目标数>放置数）
- analyze: L1 诊断指标（失败完成度、胜利剩余步数±标准差、每步消除量）
- create: 生成空白关卡模板

修复分析基础设施：
- SharedSimulationContext 补 ObstacleSystem + CellEliminator（P0: 障碍物目标不被追踪）
- FastMoveScorer 补 Obstacle 目标启发式评分 + 消除热路径 int[] 分配（P1）
- LevelAnalysisService stddev 计算加 Math.Max(0,...) 防 NaN

LevelConfig byte[] → int[]：GroundHealths/CoverHealths/ObstacleStages 改为 int[]，
消除 JSON 中的 base64 编码，关卡文件可直接人工编辑。

+++ src/Match3.Core.Tests/Simulation/RoundTripTests.cs
@@ -143,7 +143,7 @@ public class RoundTripTests
 
         // Covers
         config.Covers = new CoverType[7 * 9];
-        config.CoverHealths = new byte[7 * 9];
+        config.CoverHealths = new int[7 * 9];
         config.Covers[5] = CoverType.Cage;
         config.CoverHealths[5] = 2;
         config.Covers[15] = CoverType.Chain;
@@ -153,7 +153,7 @@ public class RoundTripTests
 
         // Grounds
         config.Grounds = new GroundType[7 * 9];
-        config.GroundHealths = new byte[7 * 9];
+        config.GroundHealths = new int[7 * 9];
         config.Grounds[1] = GroundType.Ice;
         config.GroundHealths[1] = 3;
         config.Grounds[8] = GroundType.Ice;
